fix: resolve resource leaks, ineffectual assignments, and lint warnings
- Fix swallowed error return in LBInfraCleaner.cleanupInfraSharedResources
- Ensure HTTP response bodies are properly closed across NSX client, inventory services, endpoint, and auth clients to prevent connection leaks
- Eliminate ineffectual assignments and redundant struct heap allocations in store and controller layers
- Remove redundant type conversions on typed constants and string values
- Suppress bodyclose in Transport.RoundTrip complying with http.RoundTripper contract

+++ pkg/util/restore_test.go
@@ -166,7 +166,7 @@ func TestCompareNSXRestore(t *testing.T) {
 						Status: &model.GlobalRestoreStatus{
 							Value: common.String(RestoreStatusSuccess),
 						},
-						RestoreEndTime: common.Int64(int64(time.Now().AddDate(0, 0, -1).UnixMilli())),
+						RestoreEndTime: common.Int64(time.Now().AddDate(0, 0, -1).UnixMilli()),
 					}, nil
 				})
 				return patches
